Changing transaction and wallet data storage from Double to BigDecimal. Customize ui.

+++ app/src/main/java/com/direpredium/currencyexchanger/view/CurrencyExchangerView.kt
@@ -3,10 +3,11 @@ package com.direpredium.currencyexchanger.view
 import com.direpredium.currencyexchanger.model.db.entity.Transaction
 import com.direpredium.currencyexchanger.model.db.entity.Wallet
 import com.direpredium.currencyexchanger.model.network.exception.CustomException
+import java.math.BigDecimal
 
 interface CurrencyExchangerView {
     fun showWallets(wallets: List<Wallet>)
-    fun showConvertedMoney(convertedMoney: Double)
+    fun showConvertedMoney(convertedMoney: BigDecimal)
     fun showConvertDialog(transaction: Transaction)
     fun showNetworkError(ex: CustomException)
     fun showConvertError(message: String)
